+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.149] - 2026-02-03
+### Added
+* `IndexOf` to `IListExtensions`
+### Changed
+* several `LayerMaskExtensions` methods to be slightly simpler
+
 ## [0.0.148] - 2026-01-29
 ### Added
 * setter to `ScriptableObjectSingleton.Instance`